+++ crates/crabe_decision/src/strategy/offensive/attacker.rs
@@ -8,6 +8,8 @@ use crate::strategy::basics::intercept;
 use crate::action::ActionWrapper;
 use crate::message::MessageData;
 use crate::strategy::Strategy;
+use crate::utils::get_best_shooting_window_bot;
+use crate::utils::get_open_shoot_window;
 use crate::utils::object_in_bot_trajectory;
 use crate::utils::KEEPER_ID;
 use crabe_framework::data::tool::ToolData;
@@ -19,62 +21,12 @@ use crabe_math::vectors::vector_from_angle;
 use crabe_math::{shape::Line, vectors::rotate_vector};
 use nalgebra::Point2;
 
-const MARGIN_SHOOTING_WINDOW: f64 = 0.01;
 
 /// The Attacker strategy is responsible for moving the robot to the ball and then try scoring a goal
 pub struct Attacker {
     /// The id of the robot to move.
     id: u8,
     messages: Vec<MessageData>,
-}
-
-/// Get the obstruct goal zone from an enemy
-/// 
-/// # Arguments
-/// - `shoot_start_position`: The position of the object that want to go into the goal
-/// - `enemy_position`: The position of the enemy
-/// - `world`: The current state of the game world
-/// 
-/// # Returns
-/// A `Line` representing the obstruct goal zone
-fn get_obstruct_goal_zone_from_enemy(shoot_start_position: &Point2<f64>, enemy_position: &Point2<f64>, world: &World) -> Option<Line> {
-    let start_pos_to_enemy = enemy_position - shoot_start_position;
-    let perp = rotate_vector(start_pos_to_enemy.normalize(), PI/2.) * (world.geometry.robot_radius + world.geometry.ball_radius + MARGIN_SHOOTING_WINDOW);
-    let ray_left_side = (enemy_position + perp) - shoot_start_position;
-    let ray_right_side = (enemy_position - perp) - shoot_start_position;
-    let line_ray_left_side = Line::new(*shoot_start_position, shoot_start_position + ray_left_side * 1000.);
-    let line_ray_right_side = Line::new(*shoot_start_position, shoot_start_position + ray_right_side * 1000.);
-    let intersection_left = line_ray_left_side.intersection_segment_line(&world.geometry.enemy_goal.line);
-    let intersection_right = line_ray_right_side.intersection_segment_line(&world.geometry.enemy_goal.line);
-    match (intersection_left, intersection_right) {
-        (Ok(left), Ok(right)) => Some(Line::new(left, right)),
-        (Ok(left), _) => Some(Line::new(left, world.geometry.enemy_goal.line.end)),
-        (_, Ok(right)) => Some(Line::new(world.geometry.enemy_goal.line.start, right)),
-        _ => None,
-    }
-}
-
-/// Get the open shoot window for the attacker
-/// 
-/// # Arguments
-/// - `shoot_start_position`: The position of the object that want to go into the goal
-/// - `world`: The current state of the game world
-/// 
-/// # Returns
-/// A vector of `Line` representing the open shoot windows
-pub fn get_open_shoot_window(shoot_start_position: &Point2<f64>, world: &World) -> Vec<Line> {
-    let mut available_targets: Vec<Line> = vec![world.geometry.enemy_goal.line];
-    for enemy in world.enemies_bot.values() {
-        if let Some(line) = get_obstruct_goal_zone_from_enemy(shoot_start_position, &enemy.pose.position.xy(), world){
-            let mut new_targets: Vec<Line> = vec![];
-            for target_line in available_targets {
-                let targets = target_line.cut_off_segment(&line);
-                new_targets.extend(targets);
-            }
-            available_targets = new_targets;
-        }
-    }
-    return available_targets;
 }
 
 impl Attacker {
@@ -91,21 +43,7 @@ impl Attacker {
             return shoot(robot, &ball, &world.geometry.enemy_goal.line.center(), world);
         }
         let robot_position = robot.pose.position;
-        let mut closest_ally: Option<&Robot<AllyInfo>> = None;
-        let mut max_window_length = 0.;
-        for ally in allies_in_positive_x {
-            if object_in_bot_trajectory(world, robot.id, ally.pose.position, false, false, true).len() > 0 {
-                continue;
-            }
-            let shoot_windows = get_open_shoot_window(&ally.pose.position, world);
-            let total_length = shoot_windows.iter().fold(0., |acc, line| acc + line.norm());
-            if total_length <= max_window_length {
-                continue;
-            }
-            max_window_length = total_length;
-            closest_ally = Some(ally);
-        }
-        
+        let closest_ally: Option<&Robot<AllyInfo>> = get_best_shooting_window_bot(&allies_in_positive_x, world);
         match closest_ally {
             Some(ally) => {
                 let robot_to_ally = (ally.pose.position - robot_position).normalize();
